alert

+++ lib/infrastructure/repository/follow/repo.dart
@@ -1,6 +1,5 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:quick_glimpse/application/bottm_nav_bloc/bottom_nav_bloc.dart';
-import 'package:quick_glimpse/domain/models/random_user/model.dart';
 
 class FollowRepo {
   Future<void> follow(String userUid) async {
@@ -15,20 +14,16 @@ class FollowRepo {
   }
 
   Future<void> increseFollow(String userUid, bool decrease) async {
-    try {
-      final querySnapshot = await FirebaseFirestore.instance
-          .collection('Profile')
-          .where('uid', isEqualTo: userUid)
-          .get();
+    final querySnapshot = await FirebaseFirestore.instance
+        .collection('Profile')
+        .where('uid', isEqualTo: userUid)
+        .get();
 
-      final profileDoc = querySnapshot.docs.first;
-      DocumentReference profileRef = profileDoc.reference;
-      decrease
-          ? await profileRef.update({'followers': FieldValue.increment(-1)})
-          : await profileRef.update({'followers': FieldValue.increment(1)});
-    } catch (e) {
-      print('someerror on update');
-    }
+    final profileDoc = querySnapshot.docs.first;
+    DocumentReference profileRef = profileDoc.reference;
+    decrease
+        ? await profileRef.update({'followers': FieldValue.increment(-1)})
+        : await profileRef.update({'followers': FieldValue.increment(1)});
   }
 
   Future<List> checkFollow() async {
@@ -50,10 +45,12 @@ class FollowRepo {
     increseFollow(userUid, true);
     final querySnapshot = await FirebaseFirestore.instance
         .collection('Follow')
-        .where('followed_user', isEqualTo: userUid)
+        .where('follower', isEqualTo: users?.uid)
         .get();
     for (var element in querySnapshot.docs) {
-      element.reference.delete();
+      if (element['followed_user'] == userUid) {
+        element.reference.delete();
+      }
     }
     //this is the correct code change the
     // if (querySnapshot.docs.isNotEmpty) {
@@ -70,11 +67,11 @@ class FollowRepo {
           .where('uid', isEqualTo: email)
           .get();
 
-      querySnapshot.docs.forEach((element) {
+      for (var element in querySnapshot.docs) {
         final followers = element['followers'];
 
         randomuser = followers;
-      });
+      }
       return randomuser;
     } catch (e) {
       return 0;
